pyproject.toml
[project]
name = "defenseclaw"
version = "0.1.0"
description = "Enterprise governance layer for OpenClaw — secures agentic AI deployments"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "Apache-2.0" }
dependencies = [
    "click>=8.1",
    "pyyaml>=6.0",
    "requests>=2.33",
    "rich>=13.0",
    "cisco-ai-skill-scanner>=2.0",
    "cisco-ai-mcp-scanner>=1.0; python_version>='3.11'",
]

[project.optional-dependencies]
tui = ["textual>=0.50"]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.4",
]

[project.scripts]
defenseclaw = "defenseclaw.main:cli"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["cli"]

[tool.uv]
default-groups = ["dev"]

[tool.ruff]
target-version = "py310"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
testpaths = ["cli/tests"]
